requirements.txt
annotated-types==0.7.0
anyio==4.4.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
easyocr==1.7.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.0
filelock==3.16.0
firebase-admin==6.5.0
fsspec==2024.9.0
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httplib2==0.22.0
idna==3.8
imageio==2.35.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
PyJWT==2.9.0
pyparsing==3.1.4
python-bidi==0.6.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sympy==1.13.2
tifffile==2024.8.30
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
